Fix: get app dir from native part to initialize Gno's structures (#7)
This PR fixes a crash when I try to run the gnoboard app on Android and
on some iOS devices.
The Gno configuration struct `BaseOptions` needs a `dataDir` path to
create files. The default `HomeDir()` function provided in
`tm2/pkg/crypto/keys/client/common.go` to get that path doesn't work on
mobile and causes a `panic`, because some environment variables are not
set on mobile.
The `HomeDir` function is called every time the
`tm2/pkg/crypto/keys/client` subfolder is imported because of the global
variable `DefaultBaseOptions` that calls `Homedir`. To avoid the call of
the `HomeDir` function that calls a `panic`, we duplicated `BaseOptions`
in **gnomobile**, so the `tm2/pkg/crypto/keys/client` is no more
imported and we provided it the app container path directory from the
Android/iOS native parts.

This PR also updates the `go.mod` to use the latest version of **Gno**
that fixes the ambiguous import of the `btcd` dependencies.

---------

Signed-off-by: D4ryl00 <d4ryl00@gmail.com>