feat: Expo module implementation (#129)
This PR implements an Expo module in the `expo` folder.

The following code is duplicated and adapted from the Gnoboard's code:
- Java/Kotlin and Swift native code
- use-gno.ts

There is an example app in `expo/example` that imports directly the
module. The README shows how to compile the app. It is useful to test
changes that we could make in the module.
When you do a `make build`, this rules compiles the Go frameworks and
makes a "local" copy of them in the `expo/android/libs` and
`expo/ios/Frameworks` folders, otherwise it doesn't compile.

The `expo/README.md` shows how to create a new app and import the
module. It is what the public may look for.

A Github action workflow is created to add a semantic release versioning
and automatically push the module on the Berty's NPM repo. It will be
debugged in future commits.

---------

Signed-off-by: D4ryl00 <d4ryl00@gmail.com>