feat: Update expo to use any activated account (#173)
PR https://github.com/gnolang/gnonative/pull/172 update the API to
support using any activated account. We need to update the expo provider
to the same API. This PR has 3 commits:

1. Update go.mod to use the latest gnokey-mobile following changes in PR
https://github.com/gnolang/gnokey-mobile/pull/9
2. In expo, npm install the latest gnolang_gnonative.bufbuild_es and
connectrpc_es to access the new API
3. Update expo/src/api/GnoNativeApi.ts to match the new API to use any
activated account, specified by address. Note that the address
parameters are optional for temporary backwards compatibility. When
dSocial and Gnokey Mobile are updated to use the new API and specify the
address, we will do a PR to require the address parameter, and to remove
the deprecated `selectAccount` and `getActiveAccount`.

---------

Signed-off-by: Jeff Thompson <jeff@thefirst.org>