chore: gnoboard: Resolve ts:check errors, put in Makefile (#182)
Similar to dSocial PR https://github.com/gnolang/dsocial/pull/132, we
want to use `npm run ts:check` in the Makefile for gnoboard, but first
we must resolve the errors. This PR has five commits:

1. Add ts:check to package.json
2. A [previous
commit](https://github.com/gnolang/gnonative/commit/b5861145b5f58a15afb6886a3c33c031ad560975)
changes gnoboard to use `KeyInfo` instead of the alias `GnoAccount` . It
seems that this was missed in one file, so we make the same change.
3. In the catch clause, we call `GRPCError` which expects `ConnectError`
or null. But the error in the catch can be any type. So add an explicit
check for `instanceof ConnectError` .
4. The definition of `NetworkList` includes `onNetworkChange`, but
ChangeNetwork.stories doesn't have it. However, `onNetworkChange` isn't
used by the storybook, so we put a noop function.
5. Now that the errors are fixed, we can update the gnoboard Makefile.
Add a target ts_check to do `npm run ts:check`. Add this as a dependency
to the node_modules target, which is used by all the others.

---------

Signed-off-by: Jeff Thompson <jeff@thefirst.org>