chore: Remove unused gnokey-mobile gRPC communication (#194)
gnokey-mobile PR https://github.com/gnolang/gnokey-mobile/pull/35
removed the unused gRPC server, since we now use deep links.
Importantly, the code to call functions like
`gnoNativeService.ClientSignTx` are removed. So we can remove these
functions in gnonative.

* In the `GnoNativeService` interface and `gnoNativeService` type,
remove `ClientGetRemote`, `ClientListKeyInfo` and `ClientSignTx` which
were only used by the Gnokey Mobile gRPC server.
* In the `Config` and `gnoNativeService` structs, remove
`useGnokeyMobile` flag and the code in api.go which does special
handling for this flag.
* In bridge.go and the related Swift and Kotlin files, remove support
for `StartGnokeyMobileService`. This allows us to remove the dependency
on `github.com/gnolang/gnokey-mobile/service`, so remove it from go.mod.
* Run `make regenerate` which updates files in api/gen/go, including
changes for the new buf version.

This PR was tested by running `make npm.pack` in the expo folder,
installing this package in gnokey-mobile, then testing with dsocial on
Android and iOS simulators.

(Note that examples/js/react-native/hello has a copies of files from the
expo folder. These are out-of-date for recent changes in the API,
including this PR. We will do a future PR to update the hello example.)

---------

Signed-off-by: Jeff Thompson <jeff@thefirst.org>