feat: update gno dependency and add EstimateGas API function (#195)
fixes https://github.com/gnolang/gnonative/issues/9

The gno's EstimateGas function takes a signed transaction as a
parameter. So we created a utility function in gnonative to temporarily
sign a transaction to estimate the gas wanted value, then update the
transaction's GasWanted field if asked by the front, with the given
security margin.

In this PR, I also removed the Buf's registry dependency for the expo
code. Instead, we copy API files locally in a vendor directory.

This PR updates our Gno dependency, which only compiles on 64-bit
architecture.

In this PR, I bumped goland to 1.23.x, so don't forget to do `asdf
install`.

---------

Signed-off-by: D4ryl00 <d4ryl00@gmail.com>
Co-authored-by: Jeff Thompson <jeff@berty.tech>