chore: goclient: Change to an active example page (#201)
The goclient example call the Render function to show an example page.
In the past, the default gnoland node had posts in the r/demo/boards
realm. But it doesn't anymore. Change to r/gnoland/pages which does have
content.

`go run . uds` shows:
```
2025/05/15 13:36:37 <main class='gno-tmpl-page'>

# Partnerships

### Fund and Grants Program

Are you a builder, tinkerer, or researcher? If you’re looking to create awesome dApps, tooling, infrastructure,
or smart contract libraries on gno.land, you can apply for a grant. The gno.land Ecosystem Fund and Grants program
provides financial contributions for individuals and teams to innovate on the platform.

...
```

Signed-off-by: Jeff Thompson <jeff@thefirst.org>