fix: Use github.com/gnolang/gnonative/v4 as module path (#207)
In PR #193 of October last year, semantic release created a new v4
version because of breaking changes. So we have to update go.mod and
imports in the code base to use the new
`github.com/gnolang/gnonative/v4` module path.

* In go.mod, change to module github.com/gnolang/gnonative/v4
* In all .go files, import from github.com/gnolang/gnonative/v4
* In buf.gen.yaml, change the module to github.com/gnolang/gnonative/v4
* Run `make regenerate`

In the expo folder, I confirmed that `make npm.pack` successfully
builds.

Signed-off-by: Jeff Thompson <jeff@thefirst.org>