feat: Add MaxDeposit in MsgCall and MsgRun (#209)
PR https://github.com/gnolang/gno/pull/4281 is merged which adds
`MaxDeposit` to `MsgCall` and `MsgRun`. So we need to add it to the
gnonative gRPC API.
* Update go.mod to the latest github.com/gnolang/gno
* In gnonativetypes.go, add `MaxDeposit` to `MsgCall` and `MsgRun`
* In gnonativetypes.go `MsgRun`, also change `Send` from `string` to
`[]Coin` to be consistent with `MsgCall`. This is not a breaking change
because we don't have `run` in the Expo module API, so no code is using
it right now
* Run `make regenerate`
* In api.go, add the helper `convertCoins` and use it to simplify
`convertCallRequest`, `convertSendRequest` and `convertRunRequest`
* In api.go, support `MaxDeposit` in `convertCallRequest` and
`convertRunRequest`
* In expo/src/api, add optional `maxDeposit` in `call` and `makeCallTx`

Signed-off-by: Jeff Thompson <jeff@thefirst.org>