feat: bump buf version and generated files (#212)
This PR updates the buf version for the generated API files. This
version generates files that now depend on
`@bufbuild/protobuf/codegenv2` instead of codegenv1.

It also removes unused old `api/package.json` and co files in the api
directory to clean it. They were added in the begining of the project
for some (csharp or desktop?) example apps.

---------

Signed-off-by: D4ryl00 <d4ryl00@gmail.com>